requirements.txt
beanie
discord-py-interactions
githubkit
motor[srv]
python-dotenv==0.20.0
pyyaml
rapidfuzz
tansy
unidiff
